Implemented Controller of Card

+++ src/main/java/dev/eernandorezende/credit_card_api/application/responses/CardResponse.java
@@ -1,4 +1,4 @@
 package dev.eernandorezende.credit_card_api.application.responses;
 
-public record CardResponse(Integer id,  String CardName, String number, String holderName, String validThru, Short cvc) {
+public record CardResponse(Long id,  String CardName, String number, String holderName, String validThru, Short cvc) {
 }
